Make provence source explicit; see Cheri C/C++ guide Section 6.2 Ambiguous Provenance.

+++ src/execution/isolate.h
@@ -1206,7 +1206,7 @@ class V8_EXPORT_PRIVATE Isolate final : private HiddenFactory {
   Address isolate_root() const { return isolate_data()->isolate_root(); }
   static size_t isolate_root_bias() {
 #if defined(__CHERI_PURE_CAPABILITY__)
-    return OFFSET_OF(Isolate, isolate_data_) + IsolateData::kIsolateRootBias;
+    return OFFSET_OF(Isolate, isolate_data_) + (size_t) IsolateData::kIsolateRootBias;
 #else
     return OFFSET_OF(Isolate, isolate_data_) + IsolateData::kIsolateRootBias;
 #endif
